requirements.txt
boto3~=1.35.36
python-dotenv~=1.0.1
anthropic~=0.34.2
openai~=1.51.2
geopy~=2.4.1
botocore~=1.35.32
pytest
groq
bs4